feat(experiments): 支持多GPU并发调度独立实验
- run_result3_experiments.sh 和 run_result4_experiments.sh 脚本新增 TRAIN_GPU_IDS 和 TRAIN_MAX_CONCURRENT 环境变量支持
- 实现根据可用 GPU 列表和最大并发数动态分配训练任务和调度 GPU 资源
- 引入训练配置成本预估函数并据此对配置运行顺序进行排序
- 训练任务支持异步并发执行，自动管理进程和资源释放
- 更新硬件配置相关脚本，默认支持双卡 24GB GPU + 24 核 CPU + 64GB 内存服务器档位
- 评估过程中统一将半精度张量转换为 float32，保证 numpy/json 兼容性与指标计算准确性

+++ src/ship_motion/evaluate.py
@@ -321,9 +321,11 @@ def evaluate_model_with_predictions(
 
             # 预测需要先反标准化后再统计最终指标；
             # 真实标签直接使用数据集已经提供好的 y_raw。
-            preds.append(scaler.inverse_y_tensor(y_hat).detach().cpu())
-            trues.append(batch["y_raw"].detach().cpu())
-            last_states.append(batch["last_state_raw"].detach().cpu())
+            # bf16/fp16 在后续 numpy/JSON/npz 流程里兼容性较差，
+            # 所以评估缓存统一转成 float32，再做指标统计与结果保存。
+            preds.append(scaler.inverse_y_tensor(y_hat).detach().float().cpu())
+            trues.append(batch["y_raw"].detach().float().cpu())
+            last_states.append(batch["last_state_raw"].detach().float().cpu())
         if progress_bar is not None:
             progress_bar.close()
